Added SSDT for PluginType=1
PluginType=1 on CP00 is important (and absolutely critical) for getting proper power management on hackintosh systems.

This ensures that all the correct frequency vector values are loaded so the CPUs can turbo to higher frequencies under load correctly as well as step down to lower frequencies in idle. Also aids AVX workloads.